refactor: extract package for sub validator
