+++ src/main/java/vendingmachine/reader/InputMoneyReader.java
@@ -1,6 +1,6 @@
 package vendingmachine.reader;
 
-import vendingmachine.reader.validator.NumberFormatValidator;
+import vendingmachine.reader.validator.number.NumberFormatValidator;
 import vendingmachine.reader.validator.Validator;
 
 public class InputMoneyReader extends Reader<Integer>{
