+++ src/main/java/vendingmachine/reader/validator/number/NumberFormatValidator.java
@@ -1,6 +1,8 @@
-package vendingmachine.reader.validator;
+package vendingmachine.reader.validator.number;
 
-public class NumberFormatValidator implements Validator{
+import vendingmachine.reader.validator.Validator;
+
+public class NumberFormatValidator implements Validator {
 	@Override
 	public boolean validate(String value) {
 		return value.matches("[0-9]*");
